Added value "Left and right" element 'Laterality' and updated its Comment to reflect that. Also reomoved existing first sentence in Comment.
Major update to Use and Misuse, due to addition of 'Right and left' to 'Laterality data element, including guidance about how to use it appropriately.

This revision was taken from the trunk to preserve the Catalan and Spanish translations, where possible and which had been accidentally removed in previous versions on this branch. Also included added/updated Contributors from previous version of this branch.

Editorial refinements to Use and Misuse

Translated archetype from English to Norwegian Bokmål.